setup unit test for feature matrix.

+++ scoutmaster/src/test/java/au/org/scoutmaster/DatabaseProvider.java
@@ -5,14 +5,14 @@ import java.sql.DriverManager;
 import java.sql.SQLException;
 import java.sql.Statement;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.junit.After;
+
 import liquibase.Liquibase;
 import liquibase.database.jvm.JdbcConnection;
 import liquibase.exception.LiquibaseException;
 import liquibase.resource.FileSystemResourceAccessor;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.junit.After;
 
 public class DatabaseProvider
 {
@@ -23,12 +23,19 @@ public class DatabaseProvider
 	private static String username;
 	private static String password;
 
-	public static void initDatabaseProvider()
+	public static void initDatabaseProvider(boolean rebuilddb) throws SQLException
 	{
 
 		DatabaseProvider.connectionString = "jdbc:mysql://localhost?sessionVariables=storage_engine=InnoDB";
-		DatabaseProvider.username = "scoutmaster";
+		DatabaseProvider.username = "scoutmastertest";
 		DatabaseProvider.password = "master$4scout";
+
+		logger.warn(
+				"To run database tests you must have an account on your local db user:{}, password: {} with sufficient permission to drop/create a db.",
+				DatabaseProvider.username, DatabaseProvider.password);
+		logger.warn(
+				"To do so run the following grant:\n GRANT ALL PRIVILEGES ON *.* TO '{}'@'localhost' IDENTIFIED BY '{}' WITH GRANT OPTION",
+				DatabaseProvider.username, DatabaseProvider.password);
 
 		try
 		{
@@ -36,14 +43,17 @@ public class DatabaseProvider
 			final Connection conn = DatabaseProvider.getConnection();
 
 			final Statement stmt = conn.createStatement();
+			if (rebuilddb)
+			{
 
-			String sql = "DROP DATABASE if exists " + DatabaseProvider.SCOUTMASTERTEST_DB_NAME;
-			stmt.executeUpdate(sql);
-			DatabaseProvider.logger.info("Database dropped successfully...");
+				String sql = "DROP DATABASE if exists " + DatabaseProvider.SCOUTMASTERTEST_DB_NAME;
+				stmt.executeUpdate(sql);
+				DatabaseProvider.logger.info("Database dropped successfully...");
 
-			sql = "CREATE DATABASE " + DatabaseProvider.SCOUTMASTERTEST_DB_NAME;
-			stmt.executeUpdate(sql);
-			DatabaseProvider.logger.info("Database created successfully...");
+				sql = "CREATE DATABASE " + DatabaseProvider.SCOUTMASTERTEST_DB_NAME;
+				stmt.executeUpdate(sql);
+				DatabaseProvider.logger.info("Database created successfully...");
+			}
 
 			// Now the connection string to include the db name (can't include
 			// it to start with as it shouldn't exists as yet.
@@ -53,6 +63,7 @@ public class DatabaseProvider
 		catch (final SQLException e)
 		{
 			DatabaseProvider.logger.error(e, e);
+			throw e;
 		}
 
 	}
